Refactor websockets into socketService. Emit message from kafka consumer
